backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
bcrypt==4.0.1

# Email
fastapi-mail==1.4.1
jinja2==3.1.3

# Document Generation
reportlab==4.0.9
python-docx==1.1.0
markdown==3.5.2

# Integrations
slack-sdk==3.26.2
jira==3.6.0
PyGithub==2.1.1

# Image handling
Pillow==10.2.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
